Add linux-wpt-fyi-rel to list of trybots.
The goal here is to add a manually-triggered/optional tryjob to the CQ
that will run the WPT test suite.

A resource request was filed in crbug.com/1031562 and it was confirmed
there is enough capacity for this in the existing GCE pool.

There is already an existing waterfall builder at
https://ci.chromium.org/p/chromium/builders/ci/linux-wpt-fyi-rel and the new
tryjob is intended to mirror this builder and tester setup.

Bug: 937369
Change-Id: Iaaef8b36c3097b6368e64031a8bfdb1437a6dcf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1960691
Reviewed-by: Aaron Gable <agable@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Luke Z <lpz@chromium.org>