[Flake] Parse monorail_issue from the response
A new requirement was added that the flake service and recipe should
only skip retrying flaky failures that have bugs filed for them.

The logic has been enforced at the service side, and a 'monorail_issue'
field was added to the response in:
https://chromium-review.googlesource.com/c/infra/infra/+/1966173

This CL parses the monorail issue from the response and will surface
them as part of the step text.

Bug: 1033063
Change-Id: I821e32393bc61b0530e9dbf714b6b9c9a7d8203b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1972056
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>