[Flake] Fix unrecoverable deterministic failures
Before this change, the unrecoverable failures that caused the build to
fail is determined as the failures of "with patch" step, which is
incorrect becuase some of the failures can be ignored by
"retry shards with patch" and "without patch" steps.

This CL fixes the problem to avoid confusions to the users.

Bug: 838735
Change-Id: I618939a5ea3f2de625e4de7b8820e983106c6d0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1974672
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>