[code coverage] Workaround for isolateserver 404.
When isolate loses a file it causes `swarming collect` to report the
entire shard as 'null' in chromium_swarming.summary. This then causes
the entire build to stop because the chromium_swarming module treats
this as a catastrophic failure.

Since code coverage builds are encountering this issue very frequently
we would rather just mark the shard as failed and allow it to be
retried.

Bug: 1034002
Change-Id: Id2b6ca666b0bc246cbdfdcfd412375e415764f88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1977222
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Sajjad Mirza <sajjadm@chromium.org>