[code coverage] Fix a binary and Emit a step if a binary does not exist
Originally, when a binary doesn't exist, generate_coverage_metadata
step would fail, but the error messages are very implicit that it's
hard to identify exactly which test target is causing the problem.

This CL fixes a test target that doesn't exsit in the binary mapping.

this CL also emits a step to help debugging situations like this in the
future.

Bug: 1036392
Change-Id: I80f24d7100412dd4c9c6724c375d2a4b2a5be8cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1979265
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Sajjad Mirza <sajjadm@chromium.org>