[Flake] Always emit ignoring step if applicable
Previously, an ignoring step was only emitted if all the failed tests
within a test suite are known to be flaky tests, and it could result in
confusion when part of a test suite's failed tests are exonerated.

This CL changes the behavior to always emit a ignoring step as long as
at least one test is known to be flaky and exonerated.

Change-Id: I22e6dfa22989162e8950468d343090dda91fd687
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1985530
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>