Add BotSpec for defining the details of builders in chromium_tests.
BotSpec can replace the dictionaries currently used for defining builders
and provides the following benefits:

* Validation at definition time: If someone tries to define a builder that
doesn't make sense, they'll get an exception at the time the module is
imported rather than when it's used by the chromium_tests module, which may
or may not be helpfully exposed depending on how the test is specified.

* Documentation: the class definition provides a logical spot to document
the fields of the builder definitions.

* Central location for defaults: Currently BotSpec maintains backwards
compatibility by providing a mapping interface similar to an equivalently
specified dictionary. None-values do not appear in the mapping
representation, so existing code that uses get to provide defaults will
continue to work as expected. Eventually, consumer code can be updated to
access the attribute using member access and defaults can be specified once
in the class instead.

Change-Id: I3df01d453e76f007a2ca0f79bdbbf0f2fca57833
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1986290
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>