tricium_clang_tidy: ignore pnacl compile_commands
Looks like pnacl has a tricky wrapper that does a lot of interesting
tricks with compiler flags, including adding `-isystem`s, etc. This
leads to user-visible breakages, as noted in the bug.

Given pnacl's current state, it doesn't seem worth it to support tidying
it, so this CL has us not do that.

Bug: 1041079
Change-Id: Ic9c0781237526802bcbbcb9ce4be2c61e8f32133
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1999744
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>