[Flake] Expose all failures via logs
Currently, if there are too many test failures in the
'test results summary', it only shows a portion of them, and others are
omitted, and there is no way to see all of them.

To address the issue, this CL to expose all of them via logdog links.

Bug: 1042320
Change-Id: I6d87e7b30e3ca34909b9a5eac1f32b0198e3b639
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2005065
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>