tricium_clang_tidy: suppress errors if they're all we have
This CL suppresses all tidy output for a given file if all we have to
say is "this file literally can't compile somehow," (with associated
errors on lines about why it can't compile).

There are a few reasons I think this is probably the best choice for UX:

* Since we run on a vanilla Linux build config, these errors are
extremely likely to show up in the pre-CQ or CQ anyway. We may get these
errors to the dev _faster_, but they're more or less just going to
repeat what the CQ has to say.

* Errors we find might be due to a broken tree (crbug.com/1041906). In
this case, the CQ may transparently retry the change, so diags we
produce might just add unnecessary confusion.

* Tricium is generally silent unless it has something to say, so "hey,
btw, diags might be incorrect/incomplete here" probably isn't great to
say if we don't have any diags to speak of for the given file.

* If a broken CL keeps us from finding something to complain about, a
later patch-set will be diagnosable. If a broken tree keeps us from
finding something to complain about, the CL being inspected may land
silently without diagnostics. Since tidy is 100% advisory anyway, and
the best we can provide in that case is "we might not have run 100% on
your file," this seems like the right trade-off.

Bug: 1041906
Change-Id: I935dcaef67f2c668878d87dfdcd68831fbf7f566
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2001917
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>