Change --gold_key arg to use "javascript_runtime" in pdfium.py
Currently the unique --gold_key arg is passed a string of key-value
pairs describing the configurations of the tests. A portion of that
string currently reads "javascript {enabled,disabled}" to signify
whether Javascript is enabled during runtime.

In preparation for adding tests that pass --disable-xfa, we would
need to have a portion of that string signifying whether XFA is
enabled during runtime. "xfa {enabled,disabled}" would clash with the
existing key describing the build configuration of the tests. The key
would consequently need to be named "xfa_runtime".

For consistency, change "javascript" to "javascript_runtime". Set the
value of the key to false if V8 is not built in.

Change-Id: I5f20854094d4fc3b83fe8e51eeb8d9c8be036bf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2026748
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>