Skip the CheckOwners presubmit check for changes to release branches
Most changes uploaded to release branches are cherry-picks of already-
reviewed code which has previously landed on master. This change makes
the chromium presubmit recipe look at what branch the change-under-test
is targeting, and if that branch is a release branch, tells the
underlying presubmit script to skip the CheckOwners canned check.

The conditions under which the CheckOwners check are skipped should
be refined, but this is a broad first brush at making it so that folks
cherry-picking CLs to release branches don't have to use TBR or get
their change re-reviewed.

This requires https://crrev.com/c/2018224 to be submitted first.

Bug: 1034114
Change-Id: I4ae11ba6ce8afdebabb752f914d73c704c4b226d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2017888
Commit-Queue: Aaron Gable <agable@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Anthony Polito <apolito@google.com>