Add normalize operation to BotSpec.
This ensures that all specs passed to chromium_tests must be either
class-based or are dicts that can be converted to the basic BotSpec
type. This in turn ensures that all specs benefit from the validation
provided by BotSpec and enables the code that consumes the specs to be
switched to attribute access rather than dictionary access.

Change-Id: I663a89b527bb581004ceccf3786a3f11b1d7fea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2017887
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>