Error on empty output json
We assume elsewhere that an empty or missing output.json file
is an error. Change the collection script to also assume this and
fail the collection step. This should fix the linked bug, and
ideally won't affect much else; I'd be surprised if this does,
but it's possible.

Bug: 1050834
Change-Id: I6e0a6182ddabe56d08c643ce7663bb6c6619af78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2051207
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>