Add experimental flow for WebRTC tests in histogram mode.
This is a reland of change 2059968. See diff from PS1.

This will stand up a new step that invokes webrtc_dashboard_upload.py.
It will likely fail for now, but I've made it ignore failures so it
should not make the bots red.

There will be a new step webrtc_perf_tests_experimental_hist. I'm
running it with a restrictive gtest filter so it doesn't slow down the
bot too much.

This also makes the perf testers build the webrtc_dashboard_upload
target, since it's used by the perf upload script. I hope to be able
to move away from that and transport those over using isolate, but
let's go with this for now.

Bug: chromium:1029452
Change-Id: I9a3054272f1758e2cf89ce478cd7db6a7def368f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2059969
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Commit-Queue: Patrik Höglund <phoglund@chromium.org>