Bump xcode version for WebRTC bots to 11c29.
All WebRTC bots have been upgraded to 10.14.6, which is a prerequisite
for this XCode version [1].

This is necessary because Chromium build scripts have started passing
-D to libtool, and we need to upgrade XCode for that.

It turns out $depot_tools/osx_sdk doesn't report a correct value.
That property is meant to report the highest XCode used by the system,
but there must be some bug that prevents it from reporting the right
value. I will hardcode it instead - we're not updating it that often,
and having the value be searchable in the code is easy and quick to
understand. It took me a long time to even understand how the
$depot_tools variable even worked, or rather is supposed to work.

[1] https://en.wikipedia.org/wiki/Xcode#11.x_series

Bug: webrtc:11349
Change-Id: Ia3b33a8e3fe53f88c92a085c66c844f6fa091cc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2064981
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Commit-Queue: Patrik Höglund <phoglund@chromium.org>