Replace BotId with BotMirror.
Currently, BotId is used as sort-of-an ID value but also the
configuration of a mirroring relationship: two unequal BotIds can refer
to the same builder if they refer to different testers.

This change replaces BotId with BotMirror which takes advantage of the
BuilderId type to identify the builder and optional tester.

BotConfig will accept any of BuilderId, BotMirror or the dictionary that
corresponds to the existing BotId.

Bug: 1054043
Change-Id: I66c131b142cd8c3aef8a39c9cb14bbea4406ded5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2063692
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>