[Benchmarking] Some refactoring for chromium_tests/api.py
TL;DR: This CL refactored chromium_tests/api.py to provide more modularized APIs.
This is a preparation for adding new recipe for testers.

The recipe chromium.py is current used by all CI bots. That means any changes
in the main_waterfall_steps() will affect all the CI bots.

The current implementation of main_waterfall_steps() is a bit tricky. It is
shared by both the builder and tester. For example, the triggering workflow
executes before the test runner. With appropriate configs, we can assume that
the builder will not run tests, and the tester will not trigger anything. The
order of those two steps looks do not matter. Though, in the case that the bot
is both builder and tester, I'm not sure how steps will interact with each
other.

Here we only want to change the behavior of the perf testers, e.g., we need
to the triggering workflow to run after the test runner. Instead of updating
the current implementation, I think a safer way will be to take out the steps
needed to have a new method. Specifically:
 1. In chromium_tests/api.py, modularize the steps to make it more callable
 2. Added new recipe chrome_speed_tester to call the new APIs to perform
the steps needed, e.g., trigger the tests before we trigger the data processor.
    - Currently we only let linux-perf-fyi to run this new recipe.

With these changes, the next step will be update the test runner to let it
return the swarming task IDs in some ways, so that we can pass those IDs to the
data processor.

Bug: chromium:1045256
Change-Id: I9addccd2fad8403c6b70b60b962898c242d0be08
Recipe-Nontrivial-Roll: release_scripts
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2057753
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>