Move the backwards-compatibility mappping functionality into attr_utils.
BotSpec currently implements a Mapping type that allows for it to be
continued to be used like a dict where fields values that are not None
are exposed as an item in the mapping with the name of the field for the
key and the field's value as the value. This functionality is useful for
migrating other dictionaries to attrs-based types.

Bug: 1056228
Change-Id: I5523a578d95d29381a957a82e52208bd913b2600
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2075198
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>