Fix test cases for failing test steps.
After my last CL, the tests that were supposed to be failing were not
anymore. Fortunately it turns out my last CL did not remove the ability
for bots to fail. See

https://webrtc-review.googlesource.com/c/src/+/169442

for proof. It's just the test cases that are wrong, because they relied
on JSON we're no longer generating or looking in. Now we'll fail the
swarming step so that collect reports the error (like it's actually
working today).

Bug: chromium:1051927
Change-Id: I31647de7faffdb71d6b8db3fcd639d00ef0ebd7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2078263
Auto-Submit: Patrik Höglund <phoglund@chromium.org>
Commit-Queue: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>