Switch all default tests to pool:chromium.tests.template.
The CQ results on crrev.com/c/2080814 give me decent confidence in
the safety of this CL. Granted, the CQ doesn't test everything, but
short of running led on each and every builder, it's the best we got.

Will watch the tree after this lands and prep a revert.

Recipe-Nontrivial-Roll: build_limited_scripts_slave
Bug: 812428
Change-Id: I8bbf9b7f852d61b03b87b7d96b11fb03d78f3f2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2080636
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>