Reland "Temporarily move V8 swarming tests into chromium.tests.template pool."
This is a reland of 54a92215977c6b422fa2e5f3a24430fc6edee386

Reason for reland:
The cpython pins in the swarming task template were rolled back to
a version that supports armv6l. Additionally, a PRESUBMIT check was
put in place to avoid rolling the package pins to a version not
supported on arm (which will avoid this in the future).
See crrev.com/i/2674430.

LED build of v8_linux_rel_ng_triggered with this CL:
https://ci.chromium.org/swarming/task/4abbe4347bcd6110?server=chromium-swarm.appspot.com

LED build of v8_odroid_arm_rel_ng_triggered with this CL:
https://ci.chromium.org/swarming/task/4abbe48c40cdc210?server=chromium-swarm.appspot.com

Original change's description:
> Temporarily move V8 swarming tests into chromium.tests.template pool.
>
> The 'chromium.tests.template' pool is a temporary pool that's identical
> to 'chromium.tests' (same bots) but uses swarming task templates to
> setup the default bot environment.
>
> This moves all V8 tests into the temp pool while we work on adding
> templates to the main chromium.tests pool.
>
> Ran this on a sample v8_linux_rel_ng_triggered build via led:
> https://ci.chromium.org/swarming/task/4ab66ca42855f710?server=chromium-swarm.appspot.com
>
> Bug: 812428
> Change-Id: I211dfa2330b784dec7ed5ea882abae6bcc55036d
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2083590
> Commit-Queue: Michael Achenbach <machenbach@chromium.org>
> Reviewed-by: Michael Achenbach <machenbach@chromium.org>
> Reviewed-by: Tamer Tas <tmrts@chromium.org>

Bug: 812428
Change-Id: I51d330bb54acc5413533a336a9c4673ce927114c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2086735
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>