tricium_clang_tidy: perform incremental builds
Rather than `ninja -t clean`ing on every invocation, this CL has us
reuse existing build artifacts, if possible.

The repeated `unlink`ing is only necessary if we want to track the files
that failed to build, which we *sort of* only mostly do at the moment,
and which we *might* stop caring about going forward. So it may be that
we just remove that and simplify, depending on the resolution of
crbug.com/1059096 .

Bug: 1060168
Change-Id: I1d8a11ae5003c03a93812f92971a6030dc993e3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2096419
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>