tricium_clang_tidy: don't build phonies on our first pass
This CL has us not build 'phony' targets before we want to try to build
`all`. Previously, we would attempt to build all of the targets that
include a given src_file, which can cause a significant increase in the
number of build actions we need to do (sometimes >10x).

Given that the build of phonies here is meant to be an optimization (so
we don't have to run ninja N times), it seems like there's not much
point here until we can come up with a good heuristic. It's unclear that
coming up with one is a great use of time.

Bug: 1059096
Change-Id: I4b82e1c19013999a837085315028ce8eaf3a3821
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2097437
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>