[Benchmarking] Save and pass swarming task IDs by groups
This CL saves the swarming task IDs in the swarming api so that the recipe
can load and pass them to the result processor.

After this, we can expect the swarming task IDs from the processor side and
we can run merge task from there.

Bug: chromium:1045256
Change-Id: I14bb16edf0bba7b83605ca94da32234a7ab378b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2103049
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>