Add a DUMMY_TESTER type.
There are a number of builders defined in the recipe configuration code
that don't correspond to actual builders. These dummy builders exist
only to be specified as testers in a try mirror so that a try builder
can mirror a single builder that needs to run tests in multiple
configurations (specified in the src-side configs). The presence of
these dummy testers requires that no-op jobs be defined because they
specify the builder as their parent and so when the parent runs, it
emits triggers for these testers.

This change makes it so that these dummy testers appear as the tester
in a mirror without the tester having to specify the builder as a
parent. This makes it so that the builder does not try to trigger the
non-existent builders while still running tests for multiple
configurations.

Bug: 819899
Change-Id: Ief382b8dfa408c2f78a4300f4d20a5d471fbd662
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2099747
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>