[code coverage] Add arch argument to llvm-cov commands.
The arch argument is required for iOS binaries. Added it to all places
where llvm-cov is called and this arg is required.

This CL also moved "-instr-profile=" arg in llvm-cov commands to just
before the binaries, better confirming to the doc of this command.

Bug: 943883
Change-Id: I954738e216c8577f50a86a394602a47f1edd3e7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2113116
Commit-Queue: Zhaoyang Li <zhaoyangli@chromium.org>
Reviewed-by: Sajjad Mirza <sajjadm@chromium.org>