Skip  step if the main DEPS file has changed.
In order to avoid to have a light weight CQ for Chromium Rolls and let
issue slip into master, this CL skips the "gn analyze" step if the main
DEPS file changes.

A recent CL enabled -Wunreachable-code for Mac builds and we discovered
the issue on CI because the "gn analyze" step returned "No dep found".

It looks like "gn analyze" detects changes in the dependency graph but
it doesn't detects changes in "configs".

TBR=phoglund@chromium.org
Bug: webrtc:11262
Change-Id: Ieea1496b9738d309b1a22088830b8bca0558d5d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2124480
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Commit-Queue: Mirko Bonadei <mbonadei@chromium.org>