Always compile targets for binary size analysis.
Some targets are used to calculate the binary size impact of a CL
(with results posted right on the CL page with a Gerrit plugin).

These targets need to be built even if "gn analyze" doesn't find any
dependency.

This CL adds a workaround for this, but a better solution would be to
pull the config from [1] in order to avoid to duplicate the
configuration.

[1] - https://cs.chromium.org/chromium/build/scripts/slave/recipe_modules/webrtc/builders.py

TBR=phoglund@chromium.org
Bug: webrtc:11262
Change-Id: I8ec8497c45e422ce5c2b62e3ce0cbfd8a65fcedf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2127066
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Commit-Queue: Mirko Bonadei <mbonadei@chromium.org>