Fail blink web tests upload step when they fail
This means the upload step will show up when the test fails. This is useful
because it makes the link to the test results viewer visible when users
have the 'Non-green' filter in the UI selected.

This will eventually be subsumed by the new ResultsDB UI. That's at least
a quarter out though.

Bug: 1065904
Change-Id: I7a6a374035a848101606c40bff581297f57b9f47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2132609
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>