tricium_clang_tidy: note clang-tidy invocations that failed
Currently, we only note _build_ actions that failed on Tricium.
Unfortunately, there's a __clang_analyzer__ #define that allows code to
look different code based on whether or not we're running clang-tidy.

It's unclear that this way forward is the *best*, but it's probably
worth trying, so things like the linked bug are less confusing in the
future.

Bug: 1070795, 1073171
Change-Id: Iefa67b39f4d02c3573f687f5d5a0918cee6362c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2159034
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>