Fix get_affected_files_by_patch path delimiter
https://crrev.com/c/1993279 added a call to relpath in the function.
This meant that on windows, the path delimiter '/' was replaced by '\'.
Callers of the function expected all path delimeters to be '/', so this
silently broke callers. See the bug for more details.

This CL simply replaces '\' with '/', and updates the function docstring
to specify that '/' is always the path replacement. It's also a partial revert
of https://crrev.com/c/2186354, which was a temporary fix for this issue.

Bug: 1078931
Change-Id: Ib3276f99e0406aac95160cfeb08aeee3bd6a717f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2191145
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>