mac-perf: Do not use 'mac_toolchain' in perf
Through a process of deduction, I found that none of the builds for mac
actually use the 'mac_toolchain' addiitonal config, and that's only used
for iOS builders. I suspect including that configuration makes the
builder assume that the hermetic toolchain is not used.

Removing the special case for the mac should mean that the inheritance
defined for `chromium_perf` in this builder config will inherit from all
the defaults for `chromium` which should include the downloading/usage
of the CIPD package for the SDK.

R=agable@chromium.org, gbeaty@chromium.org, thakis@chromium.org

Bug: chromium:1077773
Change-Id: I535e4cf441007e074a7179e3a752419ae9b5665e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2193315
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>