Make reading the test spec an infra step.
It is unlikely that a failure to read the test spec represents an error
on the part of a developer. Instead, it is most likely due to an
infrastructure misconfiguration. Changing it to an infra step means we
can be alerted when there is an issue.

Change-Id: I0daba1184132b7996effb1ae118ce33a102347e5
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Recipe-Nontrivial-Roll: release_scripts
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2199357
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>