[benchmarking] Processor config uses platform from the tester
Base on the discussion on crbug.com/1078939, we will try to avoid adding
new bot type for the new processor VMs, and not to skip the config checks
between triggerer and triggeree. This CL will revert the changes from
crrev.com/c/2183230 on builder tests, and use 'android' as the platform
when creating config for android-pixel2-processor-perf-fyi.

Bug: chromium:1024256
Bug: chromium:1078939
Change-Id: I03b8e95f4ad3be0e31c011534bbad9f200733f0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2197774
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>