Add Dawn end2end tests on SwiftShader to Dawn bots
This adds a second component build of dawn_end2end_tests
and runs them with SwiftShader. SwiftShader experiences
issues when dynamically loaded into non-component builds.
The second build of dawn_end2end_tests also disables the
D3D12 backend because it is unsupported on Windows 7.
This is fine as D3D12 is not needed for SwiftShader testing.

Bug: dawn:417
Change-Id: Ib443dfdf8ed4caf5dc463628e2b15c8ba2bacd92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2202747
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>