[devtools] Add backend recipe for Devtools
The recipe compiles the backend and runs unit and integration tests on backend.

Future work will create a setup with the backend installed as an extension in Chromium. On this setup we will need to run e2e tests.

Also in the near future we need to find a solution for compiling wiht Goma RBE.

Bug: chromium:1073444
Change-Id: I84ee54de1d0b272b48129d1733ae3e9426e0700c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2195263
Commit-Queue: Liviu Rau <liviurau@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Tamer Tas <tmrts@chromium.org>