Treat it as an infra failure when a BotConfig cannot be looked up.
Such a failure is unlikely to be the result of a developer change and
instead usually represents a misconfiguration in some piece of
infrastructure code. Changing it to be an infra failure will ensure we
get paged for such changes.

Change-Id: If3fda7e8dbb16778dc1cfb102be710dc002afabf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2258232
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>