[ResultDB] Remove try..except around resultdb steps
The try..except block was added when we just started to integrate recipes with resultdb, because we wanted to make sure the integration was as smooth as possible.

After several rounds of fixes, so far most of resultdb steps pass, and the rest was failed by errors that are not directly related to resultdb. So we're confident to remove the block now.

Change-Id: I93b39abcc3e671d3387987a3b893bdf8f427d55b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2247217
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>