Change swarm_hashes name to match what chromium is doing.
Pinpoint is looking for a swarm_hashes formatted the same way as chromium.
Maybe the code should be changed on pinpoint side.
So that if the key is not found it searches for 'swarm_hashes'.

The code looking for the swarm_hashes is in _CheckBuildStatus.
third_party/catapult/dashboard/dashboard/pinpoint/models/quest/find_isolate.py

Bug: webrtc:11084
Change-Id: Id3a508b9f982159514ad115d809b1be0f61ffdf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2301911
Commit-Queue: Jeremy Leconte <jleconte@google.com>
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>