Remove no-longer-used Steps subclasses from Chromium recipes.
We no longer need to construct BlinkTest objects directly; the
blink tests are now run via IsolatedScriptTest instead.

This means that the only use of the PythonTest class is for
MiniInstallerTest, which is still used on the official bots :(.

So, this CL refactors the steps classes, deletes a bunch
of dead code, and updates some test cases accordingly to
preserve coverage where possible.

Change-Id: I69b7f70a40483b88eafe3c99b1b8b76102587354
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2289490
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>