Flip linux-rel to pass command lines to swarming directly.
This is the first CL in this migration that will actually flip
a CQ builder that mirrors a builder/tester pair (or, in this
case, two builder/tester pairs). These builder/testers should
still be doing a package_build/extract_build pass, so this won't
test the isolate-only transfer case.

Lots of expected JSON files in this CL since a lot of tests are
written to use 'Linux Tests' as a builder config :(.

Bug: 1108005
Change-Id: Ieacdd1e5f5b5b7edc16a68196324a49e9de4f060
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2314034
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>