Add ability to isolate tests but not run them, add to mac-arm64.
The EngProd folks have requested that we isolate the test binaries
produced by a mac-arm64 builder, so they can download them and run
them without needing to do builds themselves.

The mac-arm64 does not currently run tests, so this CL adds a
configuration flag to tell it to upload the isolates for any
configured tests but not try to run them.

Bug: 1106083
Change-Id: I7242bf6ce3db5c04db7c2a90aa7bf173f37ac082
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2333488
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>