Stop sending swarming command lines directly on the Chromium bots.
We've discovered that the approach we were using to generate the command
lines needed for swarming tasks won't work in at least one particular
configuration (where a builder runs on a mac and a tester runs on Linux),
so we need to adopt a different approach.

This CL reverts the commits that had landed thus far to switch builders
over, so we can start over. The original code to enable the feature is
still in the recipe, just not being used. I'm not sure yet whether I'll
build on top of that CL or revert that yet as well for the new approach.

This reverts:
- cc5dda273c45205cf20bbc7df6174dd601a81d81
- 81c8242381117747babfb70cbf11935a4342f9e5
- e5e555bbda663d8268152c8870609d810bc16aa8

Bug: 1108005
Change-Id: I1e8e4d348046b03f060de92ff939beae67cb1e8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2339775
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>