Switch to passing command lines for swarming tasks via properties.
This CL implements "option 2A" from the design doc:
https://docs.google.com/document/d/1UA-3h2Mmqr3BqfEfLFubveFJIk_52T9sBFgC7MShPJ4/edit?usp=sharing

Namely, instead of relying on being able to run MB on a
test-only builder in order to get the command lines needed for a
task, we will pass the command lines across as build properties.

This has the potential risk that the build properties will become
too large for buildbucket to handle (> 1MB, maybe?), but if that's
not a real concern this approach is somewhat simpler as well as
actually addressing a case the MB approach can't address.

Bug: 1108005
Change-Id: Iff034d766aa3ca6ee2190d95a0e0fe979f7b3e64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2314734
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>