Update the build example to refer to a fake file.
The example recipe unnecessarily refers to an actual file but the recipe
could never actually be executed because it's passing invalid flags in
one of the calls. The presence of an actual script creates unexpected
results when looking for usages of the actual script.

Change-Id: Ica9d66fcc8197956725982aed5ad50a747b0f3ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2350501
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>