Reland #2 of "Switch to passing command lines via an isolated file."
This is a reland of 59f6be1cfda531f5b1172bc9f3ec6d1828cacd1e.

This time the revert was because the builder and tester were
out of sync and the tester would try to fetch a nonexistent
file from the isolate server. This fixed version checks to
make sure we actually have a hash to download instead.

Original change's description:
> Reland "Switch to passing command lines via an isolated file."
>
> This reverts commit b4bbf936a543619f153aaf5cef64666c12ef0250.
>
> Reason for revert: relanding with fix to set the default isolate server property.
>

Bug: 1108005
Bug: 1115200
Change-Id: I909e2d7be9717c00ade1d1be2645613c2a24b669
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2350338
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>