Flip rest of chromium.{gpu,linux,mac,win} builders to swarm cmd lines.
This flips the rest of the builders on the above waterfalls that use
the chromium recipe (so, not the iOS builders on chromium.mac) to
setting the command lines for the swarming tasks to use directly
(rather than relying on getting them from the isolates).

This still leaves all of the fyi builders, some of the android
builders, some of the CrOS builders, all of the builders on the
related projects that use the chromium recipe (like V8), and
the internal builders, so there's still a ways to go.

Bug: 1108005
Change-Id: Ie27b6edfa685f2cd124d04773adebb80991b0592
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2353613
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>